Data validation

+++ JukeBoxPartyWeb/Controllers/SongController.cs
@@ -50,53 +50,41 @@ namespace JukeBoxPartyWeb.Controllers
 
             string? captcha = Request.Form["g-Recaptcha-Response"];
             var recaptchaResult = ReCaptchaValidator.IsValid(captcha);
-            if (recaptchaResult.Success)
+            if (ModelState.IsValid)
             {
-
-                if (model.Track != null)
+                if (recaptchaResult.Success)
                 {
-                    var uniqueFileName = GetUniqueFileName(model.Track.FileName);
-                    var uploads = Path.Combine(hostingEnvironment.WebRootPath, "media/music");
-                    var filePath = Path.Combine(uploads, uniqueFileName);
-                    await model.Track.CopyToAsync(new FileStream(filePath, FileMode.Create));
-                    model.URL = uniqueFileName;
-                    //result = RedirectToAction("Privacy","Home");
-                    //to do : Save uniqueFileName  to your db table   
 
-                    //try
-                    //{
-                    model.Duration *= 1000;
-                    try
+                    if (model.Track != null)
                     {
-                        await APICaller.PostSong(model);
-                        //ViewBag.Message += string.Format("<b>{0}</b> uploaded.<br />", model.Track.FileName);
-                    }
-                    catch (HttpRequestException ex)
-                    {
-                        FileInfo myfileinf = new FileInfo(filePath);
-                        if ((System.IO.File.Exists(uniqueFileName)))
+                        var uniqueFileName = GetUniqueFileName(model.Track.FileName);
+                        var uploads = Path.Combine(hostingEnvironment.WebRootPath, "media/music");
+                        var filePath = Path.Combine(uploads, uniqueFileName);
+                        await model.Track.CopyToAsync(new FileStream(filePath, FileMode.Create));
+                        model.URL = uniqueFileName;
+                        model.Duration *= 1000;
+                        try
                         {
-                            System.IO.File.Delete(uniqueFileName);
+                            await APICaller.PostSong(model);
+                            ViewBag.Message += string.Format("<b>{0}</b> uploaded.<br />", model.Track.FileName);
                         }
-                        // ViewBag.Message += ex.Message;
+                        catch (HttpRequestException ex)
+                        {
+                            FileInfo myfileinf = new FileInfo(filePath);
+                            if ((System.IO.File.Exists(uniqueFileName)))
+                            {
+                                System.IO.File.Delete(uniqueFileName);
+                            }
+                            // ViewBag.Message += ex.Message;
+                        }
+                    }
+                    foreach (var error in recaptchaResult.ErrorCodes)
+                    {
+                        ModelState.AddModelError(string.Empty, error);
                     }
                 }
-                foreach (var error in recaptchaResult.ErrorCodes)
-                {
-                    ModelState.AddModelError(string.Empty, error);
-                }
-
-                // }
-                // catch(Exception ex)
-                //{
-                //ViewBag.Message += string.Format("<b>{0}</b> not uploaded."+ex.Message+ "<br />", model.Track.FileName);
-                // }
-
-
             }
-            // to do  : Return something
             var list = await APICaller.GetGenres();
-
             ViewBag.CategoryList = ToSelectList(list);
             return View(model);
         }
